Eino testing/requirements.txt
# LLM-as-Judge Testing System for Energy Advice

# Web server
fastapi>=0.109.0
uvicorn>=0.27.0

# LLM and embeddings
openai>=1.0.0

# Vector database
chromadb>=0.4.0

# PDF processing
docling>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Validation
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Text processing
tiktoken>=0.5.0

# HTTP requests for Eino API
requests>=2.31.0
